+++ client/src/get-oc-client-script.js
@@ -1,17 +1,17 @@
 'use strict';
 
-const fs = require('fs');
-const path = require('path');
-
+const getOcClientBrowser = require('oc-client-browser');
 const TryGetCached = require('./try-get-cached');
 
-module.exports = function(cache){
-
+module.exports = function(cache) {
   const tryGetCached = new TryGetCached(cache);
 
-  return function(callback){
-    tryGetCached('scripts', 'oc-client', (cb) => {
-      fs.readFile(path.resolve(__dirname, './oc-client.min.js'), 'utf-8', cb);
-    }, callback);
+  return function(callback) {
+    tryGetCached(
+      'scripts',
+      'oc-client',
+      cb => getOcClientBrowser.getLib(cb),
+      callback
+    );
   };
 };