+++ src/utils/require-template.js
@@ -17,7 +17,14 @@ function isValidTemplate(template) {
   );
 }
 
-module.exports = function(template) {
+const getOcTemplate = (path) => {
+  if (require.cache && !!require.cache[path]) {
+    delete require.cache[path];
+  }
+  return require(path);
+};
+
+module.exports = function (template) {
   let ocTemplate;
   const localTemplate = path.join(
     __dirname,
@@ -28,23 +35,21 @@ module.exports = function(template) {
   const relativeTemplate = path.resolve('.', 'node_modules', template);
 
   try {
-    if (require.cache && !!require.cache[localTemplate]) {
-      delete require.cache[localTemplate];
-    }
-    ocTemplate = require(localTemplate);
+    ocTemplate = getOcTemplate(template);
   } catch (err) {
     try {
-      if (require.cache && !!require.cache[relativeTemplate]) {
-        delete require.cache[relativeTemplate];
-      }
-      ocTemplate = require(relativeTemplate);
+      ocTemplate = getOcTemplate(localTemplate);
     } catch (err) {
-      throw format(templateNotFound, template);
+      try {
+        ocTemplate = getOcTemplate(relativeTemplate);
+      } catch (err) {
+        throw new Error(format(templateNotFound, template));
+      }
     }
   }
 
   if (!isValidTemplate(ocTemplate)) {
-    throw format(templateNotValid, template);
+    throw new Error(format(templateNotValid, template));
   }
 
   return ocTemplate;